+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Changes since v2.2:
 
+- Move Go import path from `github.com/bitly/oauth2_proxy` to `github.com/pusher/oauth2_proxy`
 - Repository forked on 27/11/18
   - README updated to include note that this repository is forked
   - CHANGLOG created to track changes to repository from original fork